requirements.txt
altgraph==0.17.4
certifi==2023.11.17
charset-normalizer==3.3.2
cx-Freeze==6.15.11
docutils==0.20.1
idna==3.6
Kivy==2.2.1
Kivy-Garden==0.1.5
packaging==23.2
patchelf==0.17.2.1
Pillow==10.1.0
Pygments==2.17.2
pyinstaller-hooks-contrib==2023.10
requests==2.31.0
urllib3==2.1.0
screeninfo==0.8.1